Make getline work on unix

+++ build.py
@@ -2,6 +2,8 @@ import os
 import shutil
 import subprocess
 import sys
+import tarfile
+import urllib.request
 
 def exe(cmd: str) -> None:
     subprocess.call(cmd.split());
@@ -15,12 +17,17 @@ def rmdir(path: str) -> None:
 def cp(source: str, destination: str) -> None:
     shutil.copy(source, destination)
 
+def exeExists(exe: str) -> bool:
+    return shutil.which(exe) != None
+
 def main(args: list[str]) -> None:
     if args[0] == "build":
         exe("cmake --build build")
 
     if args[0] == "proj":
-        exe("cmake -S . -B build -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DCMAKE_BUILD_TYPE=Debug")
+        cmd = "cmake -S . -B build -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DCMAKE_BUILD_TYPE=Debug"
+        if exeExists("ninja"): cmd + " -G Ninja"
+        exe(cmd)
 
     if args[0] == "release":
         exe("cmake -S . -B release -DCMAKE_BUILD_TYPE=Release")
@@ -29,12 +36,25 @@ def main(args: list[str]) -> None:
     if args[0] == "install-ctools":
         main(["release"])
         mkdir("bin")
-        if os.name == "nt":
+
+        if sys.platform == "win32":
             cp("./release/Release/ctemplate.exe", "./bin/ctemplate.exe")
             cp("./release/Release/cenum.exe", "./bin/cenum.exe")
-        else:
-            cp("./release/ctemplate", "./bin/ctemplate")
-            cp("./release/cenum", "./bin/cenum")
+            return
+
+        cp("./release/ctemplate", "./bin/ctemplate")
+        cp("./release/cenum", "./bin/cenum")
+
+    if args[0] == "install-clang-tools":
+        if sys.platform != "darwin": return
+
+        mkdir("tmp")
+
+        if os.path.exists("raylib-5.5_macos.tar.gz"): return
+        urllib.request.urlretrieve("https://github.com/raysan5/raylib/releases/download/5.5/raylib-5.5_macos.tar.gz", "raylib-5.5_macos.tar.gz")
+        tar = tarfile.open("raylib-5.5_macos.tar.gz")
+        tar.extractall(filter="data")
+        tar.close()
 
     if args[0] == "clean":
         rmdir("bin")
